refactor documentmenu a bit
