+++ packages/editor/src/app/main/components/documentMenu/index.tsx
@@ -1,0 +1,3 @@
+import { DocumentMenu } from "./DocumentMenu";
+
+export default DocumentMenu;
