+++ packages/editor/src/index.tsx
@@ -1,18 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import * as reo from "react-error-overlay";
-
 import reportWebVitals from "./reportWebVitals";
-
-if (process.env.NODE_ENV === "development") {
-  // disables error overlays
-  // We make use of React Error Boundaries to catch exceptions during rendering of
-  // user-defined react components. It's annoying (and slow) to get the React error overlay
-  // while editing TypeCell cells
-  // Note that this breaks hot reloading
-  (reo as any).stopReportingRuntimeErrors();
-}
 
 console.log("Loading", window.location.href);
